+++ CHANGELOG.md
@@ -30,6 +30,7 @@ Given the `MAJOR.MINOR.PATCH` pattern, here is how we decide to increment:
 - Frontend: upgrade `gulp-sourcemaps` from `1.6.0` to `2.1.1`.
 - Modified org structure to change the Assistant Director to `Stacy Canan`.
 - Updated "Standing up for you" stats on the homepage.
+- Fixed grade ordering on job listing posts.
 
 ### Removed
 - Removed Handlebars from `package.json` and `cf_notifier.js`.